Backend: Moved /ping logic to /test/ping under routes/testRoute.js
